Add double points and challenge movement. Add scoreForm folder. Separated child components into scoreForm folder.
