+++ client/src/reducers/scoreForm-fakedata.json
@@ -1,51 +1,57 @@
 {
 	"exercises": [
 		{	"name": "Push Ups",
-			"points": "10",
+			"points": 10,
 			"value": "20",
 			"type": "interval"
 		},
+		{ "name": "HS Push Ups",
+			"points": 50,
+			"value": "30",
+			"type": "interval"
+
+		},
 		{	"name": "Pull Ups",
-			"points": "50",
-			"value": "0", 
+			"points": 50,
+			"value": "10", 
 			"type": "interval"
 		},
 		{	"name": "Air Squats",
-			"points": "5",			
+			"points": 5,			
 			"value": "0", 
 			"type": "interval"
 		},
 		{	"name": "Distance Swam",
-			"points": "3500",			
+			"points": 3500,			
 			"value": "1000",
 			"units": "Miles",
 			"type": "interval"
 		},			
 		{	"name": "Distance Ran",
-			"points": "1000",
+			"points": 1000,
 			"value": "100",
 			"units": "Miles", 
 			"type": "interval"
 		},
 		{	"name": "Distance Cycled",
-			"points": "400",
+			"points": 400,
 			"value": "370",
 			"units": "Miles",
 			"type": "interval"
 		},
 		{	"name": "Distance Rowed",
-			"points": "400",
+			"points": 400,
 			"value": "20",
 			"units": "Meters",
 			"type": "interval"
 		},								
-		{	"name": "1-Minute Planks",
-			"points": "200",
+		{	"name": "One Minute Planks",
+			"points": 200,
 			"value": "37",
 			"type": "interval"
 		},				
 		{	"name": "Low Intensity",
-			"points": "5000",
+			"points": 5000,
 			"value": {
 				"hr": "3",
 				"min": "45"
@@ -53,7 +59,7 @@
 			"type": "timer"
 		},				
 		{	"name": "High Intensity",
-			"points": "7000",
+			"points": 7000,
 			"value": {
 				"hr": "0",
 				"min": "30"
@@ -61,7 +67,7 @@
 			"type": "timer"
 		},				
 		{	"name": "Sports",
-			"points": "3500",
+			"points": 3500,
 			"value": {
 				"hr": "7",
 				"min": "0"
@@ -71,7 +77,7 @@
 	],
 	"bonuses": [
 		{	"name": "Nutrition",
-			"points": "1000",
+			"points": 1000,
 			"value": {
 				"mo": true,
 				"tu": true,
@@ -83,7 +89,7 @@
 			}
 		},
 		{	"name": "Stretching",
-			"points": "1000",
+			"points": 1000,
 			"value": {
 				"mo": true,
 				"tu": true,
@@ -95,7 +101,7 @@
 			}
 		},
 		{	"name": "Tabata",
-			"points": "1500",
+			"points": 1500,
 			"value": {
 				"mo": "",
 				"tu": "",
@@ -106,5 +112,14 @@
 				"su": true				
 			}
 		}	
-	]		
+	],
+	"challenge": {
+			"name": "Archer Push Ups",
+			"points": 20,
+			"value": "15",
+			"type": "interval"  			
+	},
+	"double": {
+		"name": "Distance Swam"
+	}			
 }
