--- client/src/styles/ScoreForm.css
@@ -1,7 +1,0 @@
-.scoreForm-form {
-	margin-top: 20px;
-}
-
-.ui.checkbox {
-	margin-left: 50px;
-}
